Minor Cleanup (#6)
* Minor improvements.

Added the BMC logo. Moved the `+` tab in the interventions part of inputs to the beginning, so it doesn't move when interventions are added.

* More minor improvements.

Improve the styling on the "about" section on the welcome page. Adjust the system images, so that they never go off-screen (until going below the minimum content width).

* Minor CSS Cleanup, section labeling